+++ demtools/__init__.py
@@ -2,7 +2,7 @@
 DEMTOOLS - A collection of tools for working with DEM (Digital Elevation Model) raster files.
 """
 
-__version__ = '0.10.5'
+__version__ = '0.11.0'
 
 __all__ = [
     'asproj',
@@ -12,6 +12,7 @@ __all__ = [
     'defnodata',
     'describe',
     'mvdem',
+    'tif2csv',
 ]
 
 from . import asproj
@@ -21,4 +22,5 @@ from . import csv2tif
 from . import defnodata
 from . import describe
 from . import mvdem
+from . import tif2csv
 from . import cli